# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Navalha App API",
    "version": "1.0.0",
    "description": ""
  },
  "host": [
    "localhost:3005"
  ],
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "Bearer Token"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Buscar usuários por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID Usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}